+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 Changelog
 =========
 
+## Unreleased
+
+### Changed
+- Craft 3 Beta 20 compatibility.
+
 ## 1.0.0 - 2017-05-02
 
 Initial release.